moving read csv
